Module ssb_validate::error

Custom error type with all possible variants for this library.

Error variants are imported into other crate modules as required. The snafu crate is used here to “assign underlying errors into domain-specific errors while adding context.” Note that the visibility attribute has been applied to Error to make the variants useable throughout the crate (see the snafu documentation on Controlling Visibility for more information). This approach deviates from the recommended usage of the snafu library but has been taken here to simplify reasoning about error-handling in this library.

Enums

Error

Type Definitions

Result