Module ssb_validate::message

Functions for validating messages in the form of KVT (key, value, timestamp).

Functions

par_validate_message_hash_chain_of_feed
Batch validates a collection of messages, all by the same author, ordered by ascending sequence number, with no missing messages.

par_validate_multi_author_message_hash_chain_of_feed
Batch validates a collection of out-of-order messages by multiple authors. No previous message checks are performed, meaning that missing messages are allowed, the collection is not expected to be ordered by ascending sequence number and the author is not expected to match between current and previous message.

par_validate_ooo_message_hash_chain_of_feed
Batch validates a collection of out-of-order messages by a single author. Checks of previous message hash and ascending sequence number are not performed, meaning that missing messages are allowed and the collection is not expected to be ordered by ascending sequence number.

validate_message_hash_chain
Check that a message is a valid message relative to the previous message.

validate_multi_author_message_hash_chain
Check that an out-of-order message is valid without checking the author.

validate_ooo_message_hash_chain
Check that an out-of-order message is valid.