Module ssb_validate::message_value

Functions for validating message values (ie. just the value without key and timestamp).

Structs

SsbMessageValue
Data type representing the value of a message object (KVT). More information concerning the data model can be found in the Metadata documentation.

Functions

message_value_common_checks
Validation checks which are common across all contexts. The check_previous argument is used to control checks for the optional previous_value and previous_key parameters.

par_validate_message_value_hash_chain_of_feed
Batch validates a collection of message values, all by the same author, ordered by ascending sequence number, with no missing messages.

validate_message_value_hash_chain
Check that a message is a valid message relative to the previous message.